Move ForageConfig to core.services
ForageConfig while, often consumed by UI components, is a standalone
data structure making the .services package a more appropriate home.
This also aligns things well with the Pos SDK because the
ForageTerminalSDK, a service, consumes the ForageConfig. And, while UI
may depend on services, services should never depend on UI facilities
(else it ought to live in .ui!!)
