+++ forage-android/src/test/java/com/joinforage/forage/android/ui/ForageConfigManagerTest.kt
@@ -1,6 +1,6 @@
 package com.joinforage.forage.android.ui
 
-import com.joinforage.forage.android.core.ui.element.ForageConfig
+import com.joinforage.forage.android.core.services.ForageConfig
 import com.joinforage.forage.android.core.ui.element.ForageConfigManager
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
