+++ sample-app/src/main/java/com/joinforage/android/example/ui/catalog/CatalogFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.joinforage.android.example.databinding.FragmentCatalogBinding
-import com.joinforage.forage.android.core.ui.element.ForageConfig
+import com.joinforage.forage.android.core.services.ForageConfig
 
 class CatalogFragment : Fragment() {
